requirements.txt
aenum==3.1.11
allure-pytest==2.10.0
assertpy==1.1
curlify==2.2.1
dacite==1.6.0
deepdiff==5.8.1
dotmap==1.3.30
Faker==14.2.0
Jinja2==3.1.2
jsonschema==4.15.0
pdfminer.six==20220524
psycopg2-binary==2.9.3
pydantic==1.10.1
PySocks==1.7.1
pytest-lazy-fixture==0.6.3
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
retrying==1.3.3
selene==2.0.0b5
selenium==4.2.0
tzdata==2022.2
urllib3==1.26.10
waiting==1.4.1
xmltodict==0.13.0